task_system/requirements.txt
numpy>=1.26.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
